Previously we were not checking fixpoint convergence in houdini solver correctly: If solving predicate A depends on solving predicate B, then whenever the solution for B changes, we should re-solve A again. But B may be a connected component itself and we may perform sub-fixpoint for it -- The solution change detection for B were based on whether the *final* iteration of the sub-fixpoint converges or not, which is wrong -- as long as *any* iteration of the sub-fixpoint doesn't converge, it should count as a solution change for the outer fixpoint.
While a several-line small fix is possible, I took the chance to rewrite
the fixpoint loop, as the original logic was rather convoluted -- it
conflated what the solution was with whether the solution changed or not. The new loop is taken from the [chaotic iteration algorithm](https://github.com/backtracking/ocamlgraph/blob/master/src/chaoticIteration.ml) implemented in `ocamlgraph`, which I presume should be more reliable than the original version which came from mostly me dog-sciencing.